initial refactor for csv/xls readers:
* move away from gnrlist (out of scope) into flatfiles.py
* move away from gnrxls (why a dedicated module) into flatfiles.py
* raise deprecation warning for old imports, but preserve compatibility
* refactor Reader classes to share common ancestors

refs #309
